requirements.txt
PySide6>=6.6
PyYAML>=6.0
watchdog>=4.0
watchdog>=4.0
